🚚 Move definition of functions that handles plugin options into the src/options directory
