+++ src/utils/index.ts
@@ -1,3 +1,5 @@
+import importCwd from 'import-cwd';
+
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export function isObject(value: unknown): value is Record<any, unknown> {
     return typeof value === 'object' && value !== null;
@@ -23,4 +25,16 @@ export function filterObj<T>(
         },
         {},
     );
+}
+
+export function loadModule(
+    moduleID: string,
+    errorCallback: (error: Error) => Error | string,
+): unknown {
+    try {
+        return importCwd(moduleID);
+    } catch (error) {
+        const err = errorCallback(error);
+        throw typeof err === 'string' ? new Error(err) : err;
+    }
 }
